requirements.txt
Django==4.1.5
django-oauth-login==0.2.2
psycopg2==2.9.5
PyGithub==1.58
Faker==16.6.1
octicons==2.1.0
markdown==3.4.1